Contractualize api interface on the client
