+++ app/src/common/actions/__tests__/article-list-test.ts
@@ -1,14 +1,15 @@
 import {loadArticleList} from '../article-list';
 import {Action} from '../../actions/index';
-import {api} from '../../../../test/api-mock';
+import {apiMockData} from '../../server/api-mock';
 
 describe('loadArticleList', () => {
-    it('should dispatch some actions', async (done) => {
+    it('should dispatch request then receive actions', done => {
         chai.expect(loadArticleList()).to.dispatch.actions([{
-            type: Action.RequestArticleList
+            type: Action.RequestArticleList,
+            isLoading: true
         }, {
             type: Action.ReceiveArticleList,
-            list: await api.loadArticleList()
+            list: apiMockData.loadArticleList
         }], done);
     });
 });
