+++ app/src/common/server/index.d.ts
@@ -1,0 +1,8 @@
+import {Article} from '../definitions/article';
+
+/** Api interface, to abstract away the actual implementation in actions. */
+export interface Api {
+
+    /** Loads all the articles */
+    loadArticleList: () => Promise<Article[]>;
+}
