--- app/test/api-mock.ts
@@ -1,8 +1,0 @@
-export const api = {
-    async loadArticleList() {
-        return [
-            {id: 1, title: 'Title', description: 'Description', content: 'Content'},
-            {id: 2, title: 'Title2', description: 'Description2', content: 'Content2'}
-        ];
-    }
-}
